+++ argilla-frontend/v1/domain/services/IFieldRepository.ts
@@ -1,0 +1,11 @@
+interface Field {
+  id: string;
+  name: string;
+  required: boolean;
+  title: string;
+  settings: any;
+}
+
+export interface IFieldRepository {
+  getFields(datasetId: string): Promise<Field[]>;
+}