+++ argilla-frontend/components/features/dataset-creation/configuration/DatasetConfigurationForm.vue
@@ -90,11 +90,10 @@
                 <DatasetConfigurationQuestion
                   v-for="question in dataset.selectedSubset.questions"
                   :key="question.name"
+                  :selectedSubset="dataset.selectedSubset"
                   :question="question"
-                  :columns="dataset.selectedSubset.columns"
                   :remove-is-allowed="true"
                   :available-types="availableQuestionTypes"
-                  @remove="dataset.selectedSubset.removeQuestion(question.name)"
                   @change-type="onTypeIsChanged(question.name, $event)"
                   @is-focused="isFocused = $event"
                 />